+++ skeleton/src/Skeleton.hs
@@ -1,6 +1,0 @@
-module Skeleton (message) where
-
-import           Cardano.Prelude
-
-message :: Text
-message = "Hello, Haskell!"